+++ src/components/CopyButton.tsx
@@ -1,88 +1,21 @@
-/** @jsxRuntime classic */
-
-/** @jsx jsx */
-import { useState } from 'react';
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-import { css, jsx } from '@emotion/react';
-import styled from '@emotion/styled';
-import copy from 'copy-to-clipboard';
-import { RiCheckLine } from 'react-icons/ri';
-
-import { Colors, Spaces } from '@/assets/theme';
-
-const Button = styled.button`
-    position: relative;
-    color: ${Colors.primary};
-    padding: ${Spaces.small};
-    background: none;
-    border: 1px solid transparent;
-    border-radius: 4px;
-
-    &:hover {
-        border-color: ${Colors.primary};
-        background-color: ${Colors.darkerBackground};
-    }
-
-    &:active {
-        background-color: white;
-    }
-`;
-
-const IconContainer = styled.div`
-    position: absolute;
-    top: 0;
-    bottom: 0;
-    right: 0;
-    left: 0;
-    background-color: white;
-    border-radius: 4px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    opacity: 0;
-    pointer-events: none;
-    transition: opacity 0.1s ease-out;
-
-    svg {
-        height: 20px;
-        width: 20px;
-    }
-`;
-
-const activeStyles = css`
-    opacity: 1;
-    pointer-events: initial;
-`;
+import { Button, useClipboard } from '@chakra-ui/react';
 
 interface CopyButtonProps {
     value: string;
-    children: React.ReactNode;
 }
 
-const CopyButton = ({
-    value,
-    children,
-    ...props
-}: CopyButtonProps): JSX.Element => {
-    const [copied, setCopied] = useState(false);
-
-    const onClick = () => {
-        copy(value);
-        setCopied(true);
-        setTimeout(() => setCopied(false), 2000);
-    };
+const CopyButton = ({ value }: CopyButtonProps): JSX.Element => {
+    const { hasCopied, onCopy } = useClipboard(value);
 
     return (
-        <Button type="button" onClick={onClick} {...props}>
-            {children}
-            <IconContainer
-                css={copied && activeStyles}
-                aria-hidden={!copied}
-                aria-label="copied!"
-            >
-                <RiCheckLine />
-            </IconContainer>
+        <Button
+            variant="ghost"
+            color="teal.500"
+            textTransform="uppercase"
+            size="xs"
+            onClick={onCopy}
+        >
+            {hasCopied ? 'Copied' : 'Copy'}
         </Button>
     );
 };